Added Ackermann Drive control interface
Besides the plain throttle/brake interface a basic ackermann drive
control interface is added.

The implemented controller is a cascaded PID controller.
The first stage controller controls the speed:
- desired input: target speed (from AckermannDrive)
- measured input:  current speed (from CARLA)
- control output: accleration delta
The accleration delta is integrated over time into a acceleration target
value, which is bound by the provided AckermannDrive desired
acceleration and the maximum acceleration values of the vehicle to be
controlled.

The second stage controller controls the acceleration:
- desired input: target acceleration value (the output of the first
stage)
- measured input: current acceleration (derived from current speed to
focus on longitudinal component, slightly average filtered)
- control output: pedal delta
The pedal delta is integrated over time into a pedal target value, which
is bound by the provided maximum acceleration values of the vehicle to
be controlled.

Finally, the pedal target value is mapped from its interval
[-max_decleration; max_acceleration] towards the actual pedal range.
The mapping translates pedal target values [-max_decleration;
brake_upper_border] to the brake pedal and pedal target values
[throttle_lower_border; max_acceleration] to the throttle pedal.
The throttle_lower_border is defined by the acceleration impedance of
the driving vehicle.
The brake_upper_border is defined by the deceleration of the vehicle on
laying off the throttle pedal.
As a result the pedal target values of the interval [brake_upper_border;
throttle_lower_border] translate to a coasting of the vehicle.

In addition, some special handling for stopping and reverse were added.

To parametrize the controller a dynamic reconfigure interface
EgoVehicleControlParameters.cfg was added, the control information is
published via EgoVehicleControlInfo.msg and a rqt setup
(EgoVehicleControlInfo.perspective, client_with_rqt.launch) is provided.
Furthermore, the settings.yaml file is extended to allow the
configuration of the vehicle.